[ADD/FIX] renamed parse-set-... to utils and added other helper functions used in panel-settings. Fixed error in README. I don't believe yarn start would work anymore
